fixed wrong list name

+++ test/kext/sys/ys/homme/kext_sys_ys_homme_test.py
@@ -176,7 +176,8 @@ class KExtSysYSHommeTest(KExtSysYSTest):
             for statefile in statefiles:
                 shutil.copyfile(os.path.join(datadir, statefile), '%s/kernel/%s'%(workdir, statefile))
 
-            shutil.copyfile(os.path.join(datadir, 'kgen_statefile.lst'), '%s/kernel/kgen_statefile.lst'%workdir)
+            if s.path.exists(os.path.join(datadir, 'kgen_statefile.lst')):
+                shutil.copyfile(os.path.join(datadir, 'kgen_statefile.lst'), '%s/kernel/kgen_statefile.lst'%workdir)
 
             result['goto'] = 'runkernel_task'
             self.set_status(result, myname, self.PASSED)
